+++ WebApplication/src/helpers/resource-helpers.js
@@ -34,7 +34,7 @@ export default {
         return title.join(" | ");
     },
 
-    replaceResourceTagsInString(template, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId) {
+    replaceResourceTagsInStringPlainText(template, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel) {
         let output = template;
         let transformedInstanceTags = {};
         if (instanceTags != null) {
@@ -56,7 +56,25 @@ export default {
         return output;
     },
 
-    replaceResourceTagsInJson(template, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId) {
+    async replaceResourceTagsInStringWithTokens(template, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel) {
+        let output = template;
+
+        const matchStudyResourceToken = output.match(/\{study-resource-token\/(.*?)\}/);
+        if (matchStudyResourceToken) {
+            const tokenType = matchStudyResourceToken[1];
+            const resourceToken = await api.createToken({ 
+                tokenType: tokenType, 
+                resourcesIds: [resourceId], 
+                level: resourceLevel, 
+                validityDuration: store.state.configuration.tokens.InstantLinksValidity
+            });
+            output = output.replace('{study-resource-token/' + tokenType + '}', resourceToken['Token']);
+        }
+
+        return this.replaceResourceTagsInStringPlainText(output, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel);
+    },
+
+    async replaceResourceTagsInJson(template, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel) {
         if (template == null) {
             return null;
         }
@@ -65,7 +83,7 @@ export default {
         for (const [k, v] of Object.entries(template)) {
             if (typeof v === 'string') {
                 if (v.indexOf('{') != -1) {
-                    output[k] = this.replaceResourceTagsInString(v, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId);
+                    output[k] = await this.replaceResourceTagsInStringWithTokens(v, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel);
                 } else {
                     output[k] = v;
                 }
@@ -74,14 +92,14 @@ export default {
                 for (const vv of v) {
                     if (typeof vv === 'string') {
                         if (vv.indexOf('{') != -1) {
-                            output[k].push(this.replaceResourceTagsInString(v, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId));
+                            output[k].push(await this.replaceResourceTagsInStringWithTokens(v, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel));
                         } else {
                             output[k].push(v);
                         }
                     }
                 }
             } else if (typeof v === 'object') {
-                output[k] = this.replaceResourceTagsInJson(v, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId);
+                output[k] = await this.replaceResourceTagsInJson(v, patientMainDicomTags, studyMainDicomTags, seriesMainDicomTags, instanceTags, resourceId, resourceLevel);
             } else {
                 output[k] = v;
             }